#59 move all silhouette files in server into a new silhouette package
