+++ api/server/app/de/htwg/zeta/server/controller/restApi/FileRestApi.scala
@@ -10,7 +10,7 @@ import com.mohiva.play.silhouette.api.actions.SecuredRequest
 import de.htwg.zeta.common.format.entity.FileFormat
 import de.htwg.zeta.common.models.entity.File
 import de.htwg.zeta.persistence.general.FileRepository
-import de.htwg.zeta.server.util.auth.ZetaEnv
+import de.htwg.zeta.server.silhouette.ZetaEnv
 import grizzled.slf4j.Logging
 import play.api.data.validation.ValidationError
 import play.api.libs.json.JsError
