+++ api/server/app/de/htwg/zeta/server/controller/restApi/FilterRestApi.scala
@@ -12,7 +12,7 @@ import de.htwg.zeta.common.models.entity.File
 import de.htwg.zeta.common.models.entity.Filter
 import de.htwg.zeta.persistence.general.FileRepository
 import de.htwg.zeta.persistence.general.FilterRepository
-import de.htwg.zeta.server.util.auth.ZetaEnv
+import de.htwg.zeta.server.silhouette.ZetaEnv
 import grizzled.slf4j.Logging
 import play.api.data.validation.ValidationError
 import play.api.libs.json.JsArray
