+++ api/server/app/de/htwg/zeta/server/controller/restApi/ModelRestApi.scala
@@ -17,7 +17,7 @@ import de.htwg.zeta.persistence.accessRestricted.AccessRestrictedGraphicalDslIns
 import de.htwg.zeta.persistence.accessRestricted.AccessRestrictedGraphicalDslRepository
 import de.htwg.zeta.server.model.modelValidator.generator.ValidatorGenerator
 import de.htwg.zeta.server.model.modelValidator.validator.ModelValidationResult
-import de.htwg.zeta.server.util.auth.ZetaEnv
+import de.htwg.zeta.server.silhouette.ZetaEnv
 import grizzled.slf4j.Logging
 import play.api.libs.json.JsArray
 import play.api.libs.json.JsError
