+++ api/server/app/de/htwg/zeta/server/controller/restApi/TimedTaskRestApi.scala
@@ -10,7 +10,7 @@ import com.mohiva.play.silhouette.api.actions.SecuredRequest
 import de.htwg.zeta.common.format.entity.TimedTaskFormat
 import de.htwg.zeta.common.models.entity.TimedTask
 import de.htwg.zeta.persistence.general.TimedTaskRepository
-import de.htwg.zeta.server.util.auth.ZetaEnv
+import de.htwg.zeta.server.silhouette.ZetaEnv
 import play.api.libs.json.JsValue
 import play.api.libs.json.Writes
 import play.api.mvc.AnyContent
