+++ api/server/app/de/htwg/zeta/server/routing/authentication/AbstractAction.scala
@@ -4,7 +4,7 @@ import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
 
 import com.mohiva.play.silhouette.api.Silhouette
-import de.htwg.zeta.server.util.auth.ZetaEnv
+import de.htwg.zeta.server.silhouette.ZetaEnv
 import grizzled.slf4j.Logging
 import play.api.i18n.Messages
 import play.api.i18n.MessagesApi
