+++ api/server/app/de/htwg/zeta/server/routing/authentication/AuthenticatedWebSocket.scala
@@ -7,7 +7,7 @@ import akka.stream.Materializer
 import com.mohiva.play.silhouette.api.HandlerResult
 import com.mohiva.play.silhouette.api.Silhouette
 import com.mohiva.play.silhouette.api.actions.SecuredRequest
-import de.htwg.zeta.server.util.auth.ZetaEnv
+import de.htwg.zeta.server.silhouette.ZetaEnv
 import play.api.mvc.AnyContent
 import play.api.mvc.Request
