+++ api/server/app/de/htwg/zeta/server/routing/authentication/UnAuthenticatedAction.scala
@@ -5,7 +5,7 @@ import scala.concurrent.Future
 
 import com.mohiva.play.silhouette.api.HandlerResult
 import com.mohiva.play.silhouette.api.Silhouette
-import de.htwg.zeta.server.util.auth.ZetaEnv
+import de.htwg.zeta.server.silhouette.ZetaEnv
 import play.api.i18n.MessagesApi
 import play.api.mvc.Result
 import play.api.mvc.Request
