+++ api/server/app/de/htwg/zeta/server/silhouette/CustomSecuredErrorHandler.scala
@@ -1,4 +1,4 @@
-package de.htwg.zeta.server.util.auth
+package de.htwg.zeta.server.silhouette
 
 import javax.inject.Inject
