+++ api/server/app/de/htwg/zeta/server/silhouette/CustomUnsecuredErrorHandler.scala
@@ -1,4 +1,4 @@
-package de.htwg.zeta.server.util.auth
+package de.htwg.zeta.server.silhouette
 
 import scala.concurrent.Future
