+++ api/server/conf/shared.conf
@@ -28,7 +28,7 @@ play.http.filters = "de.htwg.zeta.server.util.Filters"
 # ~~~~~
 play.modules.enabled += "de.htwg.zeta.persistence.PersistenceModule"
 play.modules.enabled += "de.htwg.zeta.server.module.JsonFormatModule"
-play.modules.enabled += "de.htwg.zeta.server.module.SilhouetteModule"
+play.modules.enabled += "de.htwg.zeta.server.silhouette.SilhouetteModule"
 play.modules.enabled += "play.api.libs.mailer.MailerModule"
 
 play.modules.disabled += "com.mohiva.play.silhouette.api.actions.SecuredErrorHandlerModule"
